Split test_spectrum.py files
Restructure tests in line with mismatch.py restructure (#2136)
